+++ R/paramsBase_ui.R
@@ -1,30 +1,35 @@
 quipuLayout_ui <- function(ids = paste0("sample.", 1:3)){
 
-  tagList(
-  fluidRow(
-    column(width = 6,
-           radioButtons("quipuLayout", "Layout",list("full","no text"), "full", inline = TRUE)
+  shiny::tagList(
+    shiny::fluidRow(
+      shiny::column(width = 6,
+        shiny::sliderInput("quipuWidth", label = "Chart width", min = 1500, max = 6000, value = 1500, step = 300)
+        
+      ),
+      shiny::column(width = 6,
+        shiny::selectInput("quipuChartFormat", label = "Chart format", c("screen", "png", "jpg"))
+      )
+    ),    
+  shiny::fluidRow(
+    shiny::column(width = 6,
+           shiny::radioButtons("quipuLayout", "Layout", list("full","no text"), "full", inline = TRUE),
+           shiny::conditionalPanel("input.quipuLayout == 'full'",
+                       shiny::checkboxInput("quipuShowTextBox", "Show legend textbox",value = TRUE),
+                       shiny::sliderInput("quipuTextSize", "Text size", min = 0.8, max = 5, value = .6)
            ),
-    column(width = 6,
-           conditionalPanel("input.quipuLayout == 'full'",
-                            textInput("quipuIdLabel","Identifier label","ID label")
-           )
-    )
-  ),
-  fluidRow(
-    column(width = 6,
-           conditionalPanel("input.quipuLayout == 'full'",
-            checkboxInput("quipuShowTextBox", "Show legend textbox",value = TRUE)
-           )
+           shiny::checkboxInput("quipuShowHorLines", "Show horizontal lines", value = TRUE),
+           shiny::sliderInput("quipuVerLineSize", "Vertical line size", min = 1, max = 5, value = 2)
            ),
-    column(width = 6,
-           conditionalPanel("input.quipuShowTextBox == true & input.quipuLayout == 'full'",
-                            textInput("quipuSpeciesName","Species name","Sample species"),
-                            textInput("quipuSetName","Set name","Sample set"),
-                            checkboxInput("quipuShowAccsTotal", "Show total of accesions", TRUE)
-           )
+    shiny::column(width = 6,
+                  shiny::conditionalPanel("input.quipuShowTextBox == true & input.quipuLayout == 'full'",
+                      shiny::textInput("quipuIdLabel","Identifier label","ID label"),
+                      shiny::textInput("quipuSpeciesName","Species name","Sample species"),
+                      shiny::textInput("quipuSetName","Set name","Sample set"),
+                      shiny::checkboxInput("quipuShowAccsTotal", "Show total of accesions", TRUE)
+                  )                  
     )
   )
+ 
   )
 }